add relationship between idea and user

+++ src/idea/idea.service.ts
@@ -3,50 +3,66 @@ import { Injectable, HttpException, HttpStatus } from '@nestjs/common';
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { IdeaEntity } from './idea.entity';
-import {IdeaDTO} from './idea.dto';
+import {IdeaDTO, IdeaRO} from './idea.dto';
+import { UserEntity} from '../user/user.entity';
 
 @Injectable()
 export class IdeaService {
     constructor(
-    @InjectRepository(IdeaEntity)
-    private ideaRepository: Repository<IdeaEntity>,
+        @InjectRepository(IdeaEntity)
+        private ideaRepository: Repository<IdeaEntity>,
+        @InjectRepository(UserEntity)
+        private userRepository: Repository<UserEntity>,
     ) {}
 
-    async showAllIdeas() {
-        return await this.ideaRepository.find(); 
+    private toResponseObject(idea: IdeaEntity) : IdeaRO {
+        return { ...idea, author: idea.author.toResponseObject(false) };
     }
 
-    async create(data: IdeaDTO) {
-        this.ideaRepository.create(data);
-        const idea = await this.ideaRepository.save(data);
-        return idea;
+    private ensureOwnership(idea: IdeaEntity, userId: string) {
+        if(idea.author.id !== userId) {
+            throw new HttpException('Incorrect user', HttpStatus.UNAUTHORIZED)
+        }
     }
 
-    async read(id: string) {
-        const idea = await this.ideaRepository.findOne({where: {id: id}});
+    async showAllIdeas() : Promise<IdeaRO[]> {
+        const ideas = await this.ideaRepository.find({ relations: ['author'] }); 
+        return ideas.map( (idea) => this.toResponseObject(idea) );
+    }
+
+    async create(userId: string, data: IdeaDTO) : Promise<IdeaRO> {
+        const user = await this.userRepository.findOne({ where: {id: userId} })
+        const idea = this.ideaRepository.create({ ...data, author: user });
+        await this.ideaRepository.save(idea);
+        return this.toResponseObject(idea);
+    }
+
+    async read(id: string) : Promise<IdeaRO> {
+        const idea = await this.ideaRepository.findOne({where: {id: id}, relations: ['author']});
         if(!idea) {
             throw new HttpException('Not found', HttpStatus.NOT_FOUND);
         }
-        return idea;
+        return this.toResponseObject(idea);
     }
 
-    async update(id: string, data: Partial<IdeaDTO>) {
-        let idea = await this.ideaRepository.findOne({where: {id: id}});
+    async update(id: string, userId: string,  data: Partial<IdeaDTO>) : Promise<IdeaRO> {
+        let idea = await this.ideaRepository.findOne({where: {id: id}, relations: ['author']});
         if(!idea) {
             throw new HttpException('Not found', HttpStatus.NOT_FOUND);
         }
+        this.ensureOwnership(idea, userId);
         await this.ideaRepository.update({id}, data);
-        idea = await this.ideaRepository.findOne({where: {id: id}});
-        return idea;
+        idea = await this.ideaRepository.findOne({where: {id: id}, relations: ['author']});
+        return this.toResponseObject(idea);
     }
 
-    async delete(id: string) {
-        const idea = await this.ideaRepository.findOne({where: {id: id}});
+    async delete(id: string, userId: string) {
+        const idea = await this.ideaRepository.findOne({where: {id: id}, relations: ['author']});
         if(!idea) {
             throw new HttpException('Not found', HttpStatus.NOT_FOUND);
         }
+        this.ensureOwnership(idea, userId);
         await this.ideaRepository.delete({id});
-
-        return idea;
+        return this.toResponseObject(idea);
     }
 }
